Handle rate-limit edge case (#911)
If a large numberOfExecutions value (e.g. int.MaxValue) is used with a small perTimeSpan value, the division between the two will create a number too small to be represented by a tick, ending up with a TimeSpan.Zero value.
This then created a confusing exception message saying that the user had specified TimeSpan.Zero, when they hadn't.
Improved the exception message by throwing a specific exception if onePer ends up with a zero or negative value once computed.